Sync gateway add formula
sync_gateway provides the syncing backend for couchbase mobile and is
used together with couchbase-lite-ios or android to develop syncing apps
for iOS or android.

Since the project itself is named sync_gateway I kept the name for the
formula as well.

This is a based on the PR #29736 to homebrew